Update Filter() Method example
filter()
Array's filter() method is similar to the map() method:

It is called on an array
It takes a function as an argument
It returns a new array
The difference is that the function passed to filter() is used as a test, and only items in the array that pass the test are included in the new array. Consider the following example:

const names = ['David', 'Richard', 'Veronika'];

const shortNames = names.filter(function(name) {
  return name.length < 6;
});
Just as before, let's break it down a bit! We have the starting array:

const names = ['David', 'Richard', 'Veronika'];
We call filter() on the names array and pass it a function as an argument:

names.filter(function(name) {
  return name.length < 6;
});
Again, just like with map(), the function that's passed to filter() gets called for each item in the names array. The first item (i.e., 'David') is stored in the name variable. Then the test is performed -- and this is what's doing the actual filtering. First, it checks the length of the name. If it's 6 or greater, then it's skipped (and not included in the new array!). But, if the length of the name is less than 6, then name.length < 6 returns true and the name _is_ included in the new array!

And lastly, just like with map(), the filter() method returns a new array instead of modifying the original array:

const shortNames = names.filter(function(name) {
  return name.length < 6;
});

console.log(shortNames);
// ['David']
Above, the value of shortNames is the new array: ['David']. Notice that it only has one name in it, because both 'Richard' and 'Veronika' are 6 characters or longer, and were filtered out.

This was just a brief overview of how the filter() method works. For a deeper dive, check out filter() on MDN.